pyproject.toml
[project]
name = "stackforge-ai"
version = "0.1.0"
description = "Your Autonomous Development Team"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.8",
    "langchain-google-genai>=3.1.0",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tavily-python>=0.7.13",
]
